Revamp locks and dynamic recovery.
Now locks and subprocesses are kept as lists,
allowing multiple to be present at a time. This
solves an issue where jobs that were pre-empted
during follow functions were setting the wrong
recovery flag, and couldn't recovery automatically,
because the follow function was usurping the global
lock name. Now, multiple locks are retained, and
multiple recovery flags can be set. this also sets
the stage for multiple subprocesses, should that
be added in the future