add support for shell pipes only, use psutil to monitor RSS memory, see #22 and #27
The command is devided by pipes and then each element is evaluated by check_shell if there are any other shell indicators (like >). If so, this part of the original shell pipe is run in the shell. Otherwise as python pipe which causes the whole pypiper pipeline to stop if this shell pipe element fails. Needs testing!